Clean up monitoring implementation after merge
- Removed duplicate files (MonitoringAggregate, DistributedTracer)
- Kept MetricsAggregate and simpler services from main branch
- Added missing events (AlertResolved, HealthCheckFailed, HealthCheckPassed)
- Fixed code style issues
- All tests passing
