+++ app/Domain/Monitoring/Events/AlertResolved.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace App\Domain\Monitoring\Events;
 
-use Carbon\Carbon;
 use Spatie\EventSourcing\StoredEvents\ShouldBeStored;
 
-class MonitoringSessionStartedEvent extends ShouldBeStored
+class AlertResolved extends ShouldBeStored
 {
     public function __construct(
-        public readonly string $sessionId,
-        public readonly Carbon $startedAt,
+        public readonly string $alertId,
+        public readonly string $alertName,
+        public readonly string $resolution,
         public readonly array $metadata = []
     ) {
     }
