+++ app/Domain/Monitoring/Services/HealthChecker.php
@@ -7,7 +7,6 @@ namespace App\Domain\Monitoring\Services;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Redis;
-use Illuminate\Support\Str;
 
 class HealthChecker
 {
