+++ tests/Unit/Domain/Monitoring/Services/HealthCheckerTest.php
